requirements.txt
tqdm
numpy
opencv-python

onnx==1.17.0
# onnxruntime==1.20.1
onnxruntime-gpu==1.20.1